requirements.txt
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.18.0
requests>=2.31.0
aiohttp>=3.9.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.0
